+++ app/routes/main_routes.py
@@ -1,0 +1,56 @@
+import os
+from flask import Blueprint, render_template, request, flash, redirect, url_for, jsonify, session, abort
+from flask_login import login_required, current_user
+from sqlalchemy import func
+from werkzeug.utils import secure_filename
+
+from app import db, socketio, GST
+from app.activity_logger import log_activity
+from app.forms import (
+    LocationForm,
+    ItemForm,
+    TransferForm,
+    ImportItemsForm,
+    DateRangeForm,
+    CustomerForm,
+    ProductForm,
+    ProductWithRecipeForm,
+    ProductRecipeForm,
+    InvoiceForm,
+    SignupForm,
+    LoginForm,
+    InvoiceFilterForm,
+    PurchaseOrderForm,
+    ReceiveInvoiceForm,
+    DeleteForm,
+    GLCodeForm,
+)
+from app.models import (
+    Location,
+    Item,
+    ItemUnit,
+    Transfer,
+    TransferItem,
+    Customer,
+    Product,
+    LocationStandItem,
+    Invoice,
+    InvoiceProduct,
+    ProductRecipeItem,
+    PurchaseOrder,
+    PurchaseOrderItem,
+    PurchaseInvoice,
+    PurchaseInvoiceItem,
+    PurchaseOrderItemArchive,
+    GLCode,
+)
+from datetime import datetime
+from app.forms import VendorInvoiceReportForm, ProductSalesReportForm
+
+main = Blueprint('main', __name__)
+
+@main.route('/')
+@login_required
+def home():
+    """Render the transfers dashboard."""
+    return render_template('transfers/view_transfers.html', user=current_user)
